Add Flatpak manifest
This commit adds a Flatpak manifest file for packaging moserial.
The manifest is meant for making nightly builds in CI and local builds.
It's based on the workflow used in GNOME Podcasts.

Additional work will need to be done to implement nightly builds.
Given GNOME's infrastructure, this will probably be easier using Meson.
I have left that bit out for now.
More steps may be necessary to add moserial to GNOME's Flatpak repo.

A similar manifest should be submitted to the Flathub repository.
This will make it more widely accessible for users to install.

The desktop file, icon, and AppStream metadata are patched.
Flatpak expects these files to be named with the full Flatpak name.
So moserial.png becomes org.gnome.moserial.png and etc.
This is not the ideal solution, nor is it complete.
It should be able to configure names with the optional '.Devel' suffix.
GNOME Podcasts does this with an application_id variable with Meson.
For development, the id would become org.gnome.moserial.Devel.
All file names would likewise follow suite.
This responsibility ultimately belongs to the build system, though.
The build system should be expanded to configure things appropriately.

The README file has been updated accordingly.
It now contains the instructions necessary for building the Flatpak.

This commit also had to workaround issues building the configure script.
The file gnome-autogen.sh is not present in the repository.
It is downloaded by the manifest.
The autogen.sh file does not call the script correctly.
This is also patched in the manifest
These issues should be fixed in the future.