Suffix the Flatpak as ".Devel"
This should be the default for developer builds.
Since this is handled in the Flatpak manifest this is fine.
The build system should ultimately handle this naming.
The build system should then take care to use the ".Devel" suffix.
It should not use this suffix when the Flathub repo attempts to build.
Of course, that's all to be handled sometime in the future.